First Commit
The app currently has nothing in the homepage except the floatingActionButton which leads to the Reminders Page.

On the Reminders Page. User can see space for title of the reminder. Time and Date of the reminder along with the buttons used to edit the time and date. The buttons are functional. There are also Cancel and Save buttons in the end from which only the Close button is functional.