Changed the view of Reminders
Before: Clicking the floatingActionButton to add a reminder or tapping a reminder from homepage showed the reminder in a new page.

Now: Clicking a reminder from the homepage expands a section below it showing the necessary options. Clicking the floatingActionButton adds a new Reminder to the homepage list.

Why: It looks better.