Notification Related Issues Solved
Months long problem of notifications has been fixed. What we did:

1. mainIsolate in homepage file was shown as active even when app was terminated.
Ans. We were not closing the isolate in dispose method.

2. The notification action was getting stopped from the second time.
Ans. We were creating the NotificationIsolate to do ping-pong check with mainIsolate everytime after a button was clicked. The registered name was getting duplicated, we are now removing name when closing the isolate.

3. More stuff I don't remember.