Added Auto-Stop in BG Service
The background service now stops on its own when the application is terminated/detached and there are no reminders available.

I was also trying to make it so it stops right after app termination if no reminders are set, but I was getting some complications. For some reason, the mainIsolate was shown not active the moment I put the app in paused state by bringing up app history section in the device. Hence, cancelled it for now.