Fix: bgIsolate null issue
Issue: Whenever bgIsolate sent a ping message to check if mainIsolate is active or not, mainIsolate would send a oong message. But bgIsolate was null in mainIsolate/homepage.

Solution: We were initialising the variable too soon. Even before the isolate was set up. It was working in debug mode for some reason but not in release. So I added a reload in the bgIsolate's null check. If null, it would lookup port again and try again. And it worked.