Fixed Issue with Notification Rationale
Issue: Atfirst, the navigatorKey.currentContext was returning null. No error was shown, just the rationale wasn't shown. And since permission wasn't given, the app was being closed.

Fix: I shifted the code to MyApp. Separated the Rationale code to a new file. Added the use of a LifeCycleObserver. And now it is showing the notificationRationale perfectly.