Fixed Recurring on 1 minute issue
Issue: Doesn't matter the interval, the recurring was done on each minute instead of with recurring interval gaps. It was because I had thought how this was gonna happen but hadn't yet added its code.

Fix: Add the code for sending repeat reminder when the repeatNumber of that reminder's division with the reminder repetition interval (something is messed up in naming, the one which is a duration) results in 0 remainder (using modulo). Only then show the repeat and done.