Fixed Recurring Notification Issue
Issue: The recurring notifications were sent earlier than expected. The problem was that the bg_service notification was getting updated earlier than the change in minute. For example, the time would change from 02:03:58 to 02:04:00 in 2 seconds but the notification update was happening early and it still showed the diff string in reference to the earlier time. To solve this, I deducted 5 seconds from actual ... Aah I don't even know how it was working. Forget it.

Fix: I made it so the bg service notification would get updated a little later. Precisely half a second later the minute changes.