Fixed Reminders with repeats getting deleted issue
Issue: When we mark reminders with repeats on "daily", "weekly" as "Done", their time would be updated only when the app is open, but when not, the reminder gets directly deleted.

Fix: The reminder.g.dart was not updated. It didn't have the value for repeatInterval attribute and hence the value was falling back to default, which was "none". Hence it got deleted. I updated the reminder.g.dart and done.