Edited incrementRepeatDuration method
Before: There was a method in the reminder class which returned the next repeatDuration of a reminder. Like, +24 Hours for "Daily", etc. But we were not using it for some reason and doing the thing again in the database file.

Now: Editing it to not return it but move the dateAndTime attribute of reminder object and up accordingly and replaced the thing we were doing in database with this.