3/3 of Adding Notification Reminder Quick Access
Goal: Tapping on a notification should open the specific reminder in app.

Currently Done: i thought that the edited time was not getting saved. It was, the homepage was not getting updated because of updating every minute, I changed it to every second. Then upon tapping a notification, it still stayed in the notification tray. I made it so all of the same groupKey were removed on tap on any of the same groupkey.