Added Default Settings
Goal : Adding a Settings section with which user can change the options in the reminder page.

Current Additions: Added default settings, which get saved in the User database in first run as the user's settings (it was previously the indiValuesDB) from which the reminder_page now takes the values from. Only the recurring setting is left out.

Next: I wanted to add the settings page from where the user settings will be changed but first have to handle the Recurring Reminder option.

Edit 1 : It seems upon fixing the bug, I removed the code which was for the data-time quick dials, (maybe others too), I would add them in the next commit.