Attached Default Quick Time set and edit functions
Goal : Adding a Settings section with which user can change the options in the reminder page.

Current Additions : Forgot to all the TimeEditButtons in last commit, so added those. And changed how to TimeSetButtons worked, and added the defaults values of those.

Next : I had forgotten about the RepeatNotification buttons. .So I'll be adding those next.