Added settings for quick access time table
Goal : Adding a Settings section with which user can change the options in the reminder page

Current Additions : Added the setting options for Quick Access Time Table. Naming is wrong here and there, I have to think them up which are different from nag, I guess. Hah. I also added the duration and datetime pickers because the defaults weren't flexible enough. Mine may not look good like them, but they are not bad looking and are flexible. And some more things maybe. It's been a while since the last commit.