Minor fix
The repeat duration modal's duration picker was not showing the correct duratin by default. Fixed.

Another problem was that it did not change with the change of selection of button, it should change. But does not. So... That is a problem but we'll have to make changes to the picker widget. Later.