Fixed inaccurate first repeat notif issue
Issue: The first repeat notification of a reminder was off by one minute. The count was starting as 1 on the time of the reminder's time.

Fix: The repeat Count of that reminder, I started it with 0 instead of 1 and it fixed it.

More: I had added a minute to the durationFormatter for fixing the shown duration at the setting section previews. But the actual problem was something else hence it was giving off values in the homescreen now. The problem was that the flexdurationpicker gave off values. Values off by 1 millisecond, so we just added it and to it and done.